requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
uvicorn==0.29.0
fastapi==0.110.1
pytz==2024.2
